+++ src/main/java/frc/robot/UDP/BallDataPacket.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.UDP;
 
 import java.nio.ByteBuffer;
