+++ src/main/java/frc/robot/UDP/DataPacket.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.UDP;
 
 /**
  * Creates interface to recieve Data packets being sent
