+++ src/main/java/frc/robot/UDP/LimelightDataPacket.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.UDP;
 
 import java.nio.ByteBuffer;
